Explicitly set character encoding for MySQL
When using MySql as a db for Gerrit, the Gerrit documentation recommends
changing the db charset to latin1, in order to allow 1000 byte keys
using the default MySQL MyISAM engine.  This can lead to spurious errors
from Gerrit regarding "Illegal mix of collations".  We can avoid this by
being explicit to the connector about which charset to use, by setting
database.url in gerrit.config.