[BUGFIX] Resolve deadlock
The ruby_block[wait_for_service] blocked recipe
execution, until Gerrit is running. However,
this was already triggered :immediately, when
the service[gerrit] is :enable'd.

As it is not yet :start'ed at that point, we
always run into the "service comes up" time out.

Instead, split up the service definition into
two parts - one that defines it an one that starts
it (plus also waits until it is running).